// src/frontend/web/react-admin/src/generated/models/faute.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Mon API
 * Documentation de l'API de votre application.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface Faute
 */
export interface Faute {
    /**
     * 
     * @type {number}
     * @memberof Faute
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof Faute
     */
    libelle: any;
    /**
     * 
     * @type {string}
     * @memberof Faute
     */
    gravite: any;
    /**
     * 
     * @type {number}
     * @memberof Faute
     */
    eleveId: any;
    /**
     * 
     * @type {number}
     * @memberof Faute
     */
    regleId: any;
    /**
     * 
     * @type {BaseModelpropertiescreatedAt}
     * @memberof Faute
     */
    createdAt?: any;
    /**
     * 
     * @type {BaseModelpropertiesupdatedAt}
     * @memberof Faute
     */
    updatedAt?: any;
    /**
     * 
     * @type {BaseModelpropertiesdeletedAt}
     * @memberof Faute
     */
    deletedAt?: any;
}