// src/frontend/web/react-admin/src/generated/models/update-faute-id-body.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Mon API
 * Documentation de l'API de votre application.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface UpdateFauteIdBody
 */
export interface UpdateFauteIdBody {
    /**
     * 
     * @type {string}
     * @memberof UpdateFauteIdBody
     */
    libelle: any;
    /**
     * 
     * @type {string}
     * @memberof UpdateFauteIdBody
     */
    gravite: any;
    /**
     * 
     * @type {number}
     * @memberof UpdateFauteIdBody
     */
    eleveId: any;
    /**
     * 
     * @type {number}
     * @memberof UpdateFauteIdBody
     */
    regleId: any;
}